Revert "move bnpl tests to test with redirects"
This reverts commit ba2f89178831fa4d73968d7f54086ede4c89b65e.
