+++ e2e/paymentsWithRedirects.test.ts
@@ -1,8 +1,6 @@
 /* eslint-disable no-undef */
 import { getElementByText, getTextInputByPlaceholder } from './helpers';
-import BasicPaymentScreen, {
-  getAllWebviewContexts,
-} from './screenObject/BasicPaymentScreen';
+import BasicPaymentScreen from './screenObject/BasicPaymentScreen';
 import homeScreen from './screenObject/HomeScreen';
 
 describe('Payment scenarios with redirects', () => {
@@ -226,32 +224,5 @@ describe('Payment scenarios with redirects', () => {
       timeout: 30000,
     });
     expect(alert.getText()).toEqual('Success');
-  });
-
-  it('Afterpay/Clearpay payment scenario', function () {
-    this.retries(2);
-
-    homeScreen.goTo('Buy now pay later');
-    homeScreen.goTo('Afterpay and Clearpay');
-
-    $('~payment-screen').waitForDisplayed({ timeout: 30000 });
-
-    BasicPaymentScreen.pay({ email: 'test@stripe.com' });
-    BasicPaymentScreen.authorize({ elementType: 'a', pause: 10000 });
-    BasicPaymentScreen.checkStatus();
-  });
-
-  it('Opens Klarna webview', function () {
-    this.retries(2);
-
-    homeScreen.goTo('Buy now pay later');
-    homeScreen.goTo('Klarna');
-
-    $('~payment-screen').waitForDisplayed({ timeout: 30000 });
-
-    BasicPaymentScreen.pay({ email: 'test@stripe.com' });
-    driver.pause(10000);
-    const webviewContexts = getAllWebviewContexts();
-    expect(webviewContexts.length).toBeGreaterThan(0);
   });
 });
